move all of legacy blitzkrieg into the core directory
